Fix libgui null deref + interpolator target changing
- Finally fixed months old intermittent crash caused by incorrect malloc size in 286:logic.c
- CUNO can now (in a hacky way) run with ASan. Check ./scripts
- Some interpolator changes to make changing targets smoother.

+++ game/src/logic.c
@@ -278,7 +278,7 @@ int supersmartAI_act(struct game_state *game)
     if (game->acted) 
         return 0;
 
-    indices = malloc( sizeof(int) * player->hand.len );
+    indices = malloc( sizeof(size_t) * player->hand.len );
     if (indices == NULL)
         return -1;
